Remove `include/grpcpp/impl/codegen/slice.h` (#31283)
