+++ include/grpcpp/ext/call_metric_recorder.h
@@ -25,8 +25,8 @@
 #include "absl/strings/string_view.h"
 #include "absl/types/optional.h"
 
-#include <grpcpp/impl/codegen/slice.h>
 #include <grpcpp/impl/codegen/sync.h>
+#include <grpcpp/support/slice.h>
 
 namespace grpc_core {
 class Arena;
